gpodder 3.7.0 - Off-Screen Bionic Chiseling + new packages + gpo + provenance
I had to rebuild everything from scratch and took the opportunity to update the README.

Adding a provenance.pl script to provide traceability with the version
of bundled software. Each package having one of its files in gPodder.app
is listed in gPodder.contents, along with its version (the md5 after the version
is used by jhbuild; it is based on source package and patches).

Some build files had been forgotten.