Allow `openssl` to be installed into `DESTDIR`
`openssl`'s configuration scripts up to version 1.1.0 use the
`INSTALL_PREFIX` variable to designate the desired staging directory,
instead of the common `DESTDIR`, which `jhbuild` expects (see
https://github.com/openssl/openssl/commit/3c65577f1af1109beb8de06420efa09188981628).
So because the files are installed not in the expected directory,
`jhbuild` fails with an error. This change adds a patch to fix this
behavior in the current `openssl` version.